Move `AccessibilePopover` logic to separate hook
