+++ packages/components/src/accessible-popover/accessible-popover/index.js
@@ -1,1 +1,2 @@
 export { default } from './component';
+export { useAccessiblePopover } from './hook';
